% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotationExample}
\alias{annotationExample}
\title{Example annotation table for demultiplexing cell barcodes}
\format{A \code{data.table} with 5 columns:
\describe{
\item{\code{project}}{Project name. Project names must be identical across samples.}
\item{\code{sample}}{Sample name}
\item{\code{lane}}{Flow cell lane number. If FASTQ files from multiple lanes are concatenated, any placeholder would be sufficient, e.g. "L001".}
\item{\code{read1_path}}{Path to read1 FASTQ file. This is the read file with UMI and cell barcode information.}
\item{\code{read2_path}}{Path to read2 FASTQ file. This read file contains genomic sequences.}
}}
\usage{
annotationExample
}
\description{
A \code{data.table} object of the annotation table for example FASTQ files.
}
\keyword{datasets}
