% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodeExample}
\alias{barcodeExample}
\title{A vector of example cell barcodes.}
\format{A vector of cell barcode sequences. Cell barcodes for this study
 (van den Brink, et al.) are of length 8.}
\usage{
barcodeExample
}
\description{
A vector containing 48  predefined cell barcodes which
 will be used for
 demultiplexing the example FASTQ files. Only the latter 48 barcodes within
 the original 96 barcodes are used here to reduce the time to run example
 codes and compile the vignette.
}
\keyword{datasets}
