% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuff_functions.R
\name{collect.qc}
\alias{collect.qc}
\title{Collect QC metrics}
\usage{
collect.qc(de, al, co, biomart.annot.dt = NA)
}
\arguments{
\item{de}{Demultiplex result. Table returned from \code{demultiplex} function.}

\item{al}{Alignment result. Table returned from \code{align.rsubread} function.}

\item{co}{Count matrix. Table returned from \code{count.umi} function.}

\item{biomart.annot.dt}{Gene information table generated from running \code{biomaRt} query on gene IDs. Must contain \emph{ensembl_gene_id} and \emph{gene_biotype}.}
}
\value{
QC metrics table
}
\description{
Collect QC metrics from demultiplexing, alignment, and counting results. Return a \code{data.table} object containing reads, transcripts, and genes information.
}
