% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_umi.R
\name{count.reads}
\alias{count.reads}
\title{Count the number of UMIs for each transcript/gene and generate the expression matrix}
\usage{
count.reads(alignment, gtf, format = "BAM", out.dir = "./Count",
  cores = max(1, parallel::detectCores()/2), output.prefix = "count",
  verbose = FALSE, logfile.prefix = format(Sys.time(),
  "\%Y\%m\%d_\%H\%M\%S"))
}
\arguments{
\item{alignment}{A character vector of the paths to input alignment files.}

\item{gtf}{Path to the gtf reference file. For generation of TxDb objects from gtf files, please refer to \code{makeTxDbFromGFF} function in \code{GenomicFeatures} package.}

\item{format}{Format of input sequence alignment files. \strong{"BAM"} or \strong{"SAM"}. Default is \strong{"BAM"}.}

\item{out.dir}{Output directory for UMI counting results. Expression table will be stored in this directory. Default is \code{"../Count"}.}

\item{cores}{Number of cores used for parallelization. Default is \code{max(1, parallel::detectCores() / 2)}.}

\item{output.prefix}{Prefix for expression table filename. Default is \code{"countUMI"}.}

\item{verbose}{Print log messages. Useful for debugging. Default to \strong{FALSE}.}

\item{logfile.prefix}{Prefix for log file. Default is current date and time in the format of \code{format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S")}.}
}
\value{
A expression matrix \code{data.table} containing the raw counts of unique \emph{UMI:transcript} pairs.
}
\description{
Count unique \emph{UMI:transcript} pairs for single cell RNA-sequencing alignment files. Write resulting table to output directory. Columns are samples (cells) and rows are transcript/gene IDs.
}
