% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleannot}
\alias{exampleannot}
\title{Example annotation table}
\format{A \code{data.table} with 6 columns:
\describe{
\item{\code{project}}{Project name}
\item{\code{id}}{Sample ID}
\item{\code{num}}{Sample number based on the order that samples are listed during fastq generation}
\item{\code{lane}}{Lane number}
\item{\code{read}}{The read. R1 means Read 1 and R2 means Read 2}
\item{\code{dir}}{Directory to the fastq file}
}}
\usage{
exampleannot
}
\description{
A \code{data.table} object of example annotation table.
}
\keyword{datasets}
