% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplebc}
\alias{examplebc}
\title{Example barcodes table.}
\format{A vector of barcode sequences. In the example dataset (van den Brink, et al.), cell barcodes are of length 8.}
\usage{
examplebc
}
\description{
A \code{data.table} object containing 96 barcodes for demultiplexing example fastq files.
}
\keyword{datasets}
