% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplefastq}
\alias{examplefastq}
\title{Example fastq \code{ShortReadQ} objects}
\format{A list of 4 \code{ShortReadQ} \code{S4} objects.}
\usage{
examplefastq
}
\description{
A list of \code{ShortReadQ} objects of example fastq reads. Contains R1 and R2 reads
from Example_01 and R1 and R2 reads from Example_02. Reads library is generated
using CEL-SEQ2 protocol.
}
\keyword{datasets}
