% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastqExample}
\alias{fastqExample}
\title{Example \code{ShortReadQ} objects of FASTQ files.}
\format{A list of 4 \code{ShortReadQ} \code{S4} objects each containing 20,000 reads.}
\usage{
fastqExample
}
\description{
A list of \code{ShortReadQ} objects containing 4 example FASTQ files from 4 cells o a publicly available dataset (van den Brink, et al.). Each object contains 20,000 reads which are a subset of original dataset. Contains paired end read 1 (R1, UMI + cell barcodes) and read 2 (R2, genomic sequence) reads. Read library was generated using CEL-Seq protocol.
}
\keyword{datasets}
