% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuff_functions.R
\name{get.QC.table}
\alias{get.QC.table}
\title{Collect QC metrics}
\usage{
get.QC.table(de, al, co, biomart.result.dt = NA)
}
\arguments{
\item{de}{Demultiplex result. Table returned from \code{demultiplex} function.}

\item{al}{Alignment result. Table returned from \code{align.rsubread} function.}

\item{co}{Count matrix. Table returned from \code{count.umi} function.}

\item{biomart.result.dt}{Gene information table generated from running \code{biomaRt} query on gene IDs.}
}
\value{
QC table
}
\description{
Collect QC metrics from demultiplexing, alignment, and counting results. Return a \code{data.table} object containing reads, transcripts, and genes information.
}
