% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38_MT}
\alias{GRCh38_MT}
\title{GRCh38 mitochondial dna fasta sequence}
\format{An object of class \code{character} of length 278.}
\usage{
GRCh38_MT
}
\description{
An vector containing GRCh38 chrMT fasta sequence.
}
\keyword{datasets}
