% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38_MT_gtf}
\alias{GRCh38_MT_gtf}
\title{GRCh38 mitochondrial gtf file}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 144 rows and 9 columns.}
\usage{
GRCh38_MT_gtf
}
\description{
A \code{data.table} object of GRCh38 mitochondrial gtf file.
}
\keyword{datasets}
