% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCm38_MT_fa}
\alias{GRCm38_MT_fa}
\title{GRCm38 mitochondrial dna fasta sequence}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 204 rows and 1 columns.}
\usage{
GRCm38_MT_fa
}
\description{
An \code{data.table} containing GRCm38 mitochondrial (MT) DNA sequence.
}
\keyword{datasets}
