% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCm38_MT_gtf}
\alias{GRCm38_MT_gtf}
\title{GRCm38 mitochondrial gtf file}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 147 rows and 9 columns.}
\usage{
GRCm38_MT_gtf
}
\description{
A \code{data.table} object of GRCm38 mitochondrial gtf file.
}
\keyword{datasets}
