% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCm38MitochondrialGTF}
\alias{GRCm38MitochondrialGTF}
\title{GRCm38 mitochondrial gene sets (GTF) file}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 147 rows and 9 columns.}
\usage{
GRCm38MitochondrialGTF
}
\description{
A \code{data.table} object of GRCm38 mitochondrial (MT) GTF file.
}
\keyword{datasets}
