% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCellRanger.R
\name{importCellRanger}
\alias{importCellRanger}
\title{Construct SCE object from Cell Ranger output}
\usage{
importCellRanger(cellRangerDir, samples,
  cellRangerOuts = "outs/filtered_feature_bc_matrix/",
  matrixFileName = "matrix.mtx.gz",
  featuresFileName = "features.tsv.gz",
  barcodesFileName = "barcodes.tsv.gz", gzipped = TRUE)
}
\arguments{
\item{cellRangerDir}{The root directory where Cell Ranger was run. This
folder should contain sample specific folders.}

\item{samples}{A vector of sample names. Must be the same as the folder
names of the samples. The cells in the final SCE object will be ordered by
samples.}

\item{cellRangerOuts}{The intermidiate path to filtered feature count files
saved in sparce matrix format. Reference genome names might need to be
appended for CellRanger version below 3.0.0 if reads were mapped to
multiple genomes when running Cell Ranger pipeline. Default
\code{"outs/filtered_feature_bc_matrix/"}.}

\item{matrixFileName}{Filename for the Market Exchange Format (MEX) sparse
matrix file. Default \emph{matrix.mtx.gz}.}

\item{featuresFileName}{Filename for the feature annotation file. It can be
\emph{features.tsv.gz} or \emph{genes.tsv}. Default
\emph{features.tsv.gz}.}

\item{barcodesFileName}{Filename for the cell barcode list file. Default
\emph{barcodes.tsv.gz}.}

\item{gzipped}{Boolean. \code{TRUE} if the Cell Ranger output files
(barcodes.tsv, features.tsv, and matrix.mtx) were
gzip compressed. \code{FALSE} otherwise. This is true after Cell Ranger
3.0.0 update. Default \code{TRUE}.}
}
\value{
A \code{SingleCellExperiment} object containing the combined count
 matrix, the feature annotations, and the cell annotation.
}
\description{
Read the filtered barcodes, features, and matrices for all
 samples from a single run of Cell Ranger output. Combine them into one big
 \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\examples{
# Example #1
# The following filtered feature, cell, and matrix files were downloaded from
# https://support.10xgenomics.com/single-cell-gene-expression/datasets/
# 3.0.0/hgmm_1k_v3
# The top 10 hg19 & mm10 genes are included in this example.
# Only the first 20 cells are included.
sce <- importCellRanger(
    cellRangerDir = system.file("extdata", package = "scruff"),
    samples = "hgmm_1k_v3_20x20")
}
