% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot10x.R
\name{plot10x}
\alias{plot10x}
\title{Visualize 10X read alignment data}
\usage{
plot10x(tenxqcDt)
}
\arguments{
\item{tenxqcDt}{A \code{data.table} object containing the genome_reads and
gene_reads columns.}
}
\value{
ggplot object showing the number of aligned reads and reads aligned
 to an gene.
}
\description{
Read QC metrics file generated by \code{tenxqc} function and plot
 QC metrics including number of aligned reads and reads aligned to an gene.
}
