% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcplots.R
\name{qcplots}
\alias{qcplots}
\title{Visualize data quality}
\usage{
qcplots(sce)
}
\arguments{
\item{sce}{An \code{SingleCellExperiment} object returned from \code{scruff} or \code{countUMI} function.}
}
\value{
A list of \code{grobs} objects ready for plotting
}
\description{
Visualize data quality from the \code{colData} of the \code{SingleCellExperiment} object and return a list of figures in \code{arrangelist} object.
}
\examples{
\dontrun{
data(sceExample, package = "scruff")
qcplots(sceExample)
}
}
