% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletDetection.R
\name{runDoubletFinder}
\alias{runDoubletFinder}
\title{Generates a doublet score for each cell via doubletFinder}
\usage{
runDoubletFinder(sce, assayType)
}
\arguments{
\item{sce}{SingleCellExperiment object. Must contain a counts matrix}
}
\value{
SingleCellExperiment object containing the
 'doublet_finder_doublet_score'.
}
\description{
Uses doubletFinder to determine cells within the dataset
 suspected to be doublets.
}
\examples{
sce <- runDoubletFinder(sce, assayType = "rsemcounts")
}
