% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletDetection.R
\name{runEmptyDrops}
\alias{runEmptyDrops}
\title{Generates an empty drops score for each cell via EmptyDrops}
\usage{
runEmptyDrops(sce, lower, assayType)
}
\arguments{
\item{sce}{SingleCellExperiment object. Must contain a counts matrix}

\item{lower}{Number. Specifies the lower bound on the total UMI count.
All cells containing UMIs less than this will be considered empty droplets.}

\item{assayType}{Character. The name of the counts matrix within your
singleCellExperiment object that you wish to run EmptyDrops.}
}
\value{
SingleCellExperiment object containing the
 'droplet_utils_empty_drops_score'.
}
\description{
Uses EmptyDrops to determine cells within the dataset
 suspected to be doublets.
}
\examples{
sce <- runEmptyDrops(sce, lower = 100, assayType = "rsemcounts")
}
