% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletDetection.R
\name{runScranDoubletScore}
\alias{runScranDoubletScore}
\title{Generates a doublet score for each cell via doubletCells}
\usage{
runScranDoubletScore(sce, assayType)
}
\arguments{
\item{sce}{SingleCellExperiment object. Must contain a counts matrix}

\item{assayType}{Character. The name of the counts matrix within your
singleCellExperiment object that you wish to run doubletCells.}
}
\value{
SingleCellExperiment object containing the 'scran_doublet_score'.
}
\description{
Uses doubletCells from the scran package to determine
 cells within the dataset suspected to be doublets.
}
\examples{
sce <- runScranDoubletScore(sce, assayType = "rsemcounts")
}
