% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sceExample}
\alias{sceExample}
\title{Example SingleCellExperiment Object}
\format{A \code{SingleCellExperiment} object.}
\usage{
sceExample
}
\description{
An example SingleCellExperiment object containing count matrix, cell and
 gene annotations, and all QC metrics for mouse mitochonrial genes generated
 from example FASTQ reads.
}
\keyword{datasets}
