% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scruff_functions.R
\name{stepping}
\alias{stepping}
\title{Visualize aligned reads}
\usage{
stepping(bamGA, chr = "1", start = 1, end = max(BiocGenerics::end(bamGA)),
  legend = FALSE)
}
\arguments{
\item{bamGA}{A GenomicAlignment object}

\item{chr}{Chromosome. Integer or "X", "Y", "MT".}

\item{start}{Genomic coordinate of the start position.}

\item{end}{Genomic coordinate of the end position.}
}
\value{
A ggplot object of aligned reads
}
\description{
Visualize mapped reads for single cell RNA-sequencing data. Arrow represents orientation of alignment. Reads are colored by their UMI tag.
}
