% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenxBamqcplot.R
\name{tenxBamqcplot}
\alias{tenxBamqcplot}
\title{Visualize 10X read alignment data}
\usage{
tenxBamqcplot(tenxBamqcDt)
}
\arguments{
\item{tenxBamqcDt}{A data frame or \code{data.table} object containing the
genome_reads and gene_reads columns.}
}
\value{
ggplot object showing the number of aligned reads and reads aligned
 to an gene.
}
\description{
Read QC metrics file generated by \code{tenxBamqc} function and plot
 QC metrics including number of aligned reads and reads aligned to an gene.
}
\examples{
# first 5000 records in the bam file downloaded from here:
# http://sra-download.ncbi.nlm.nih.gov/srapub_files/
# SRR5167880_E18_20160930_Neurons_Sample_01.bam
# see details here:
# https://trace.ncbi.nlm.nih.gov/Traces/sra/?study=SRP096558
# and here:
# https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE93421
bamfile10x <- system.file("extdata",
    "SRR5167880_E18_20160930_Neurons_Sample_01_5000.bam",
    package = "scruff")

# library(TENxBrainData)
# library(data.table)
# tenx <- TENxBrainData()
# # get filtered barcodes for sample 01
# filteredBcIndex <- tstrsplit(colData(tenx)[, "Barcode"], "-")[[2]] == 1
# filteredBc <- colData(tenx)[filteredBcIndex, ][["Barcode"]]

filteredBc <- system.file("extdata",
    "SRR5167880_E18_20160930_Neurons_Sample_01_filtered_barcode.tsv",
    package = "scruff")
# QC results are saved to current working directory
qcDt <- tenxBamqc(bam = bamfile10x,
    experiment = "Neurons_Sample_01",
    filter = filteredBc)
qcDt

g <- tenxBamqcplot(qcDt)
g
}
