% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenxqc.R
\name{tenxqc}
\alias{tenxqc}
\title{Visualize 10X read alignment data quality}
\usage{
tenxqc(bams, reference, tags = c("NH", "CB", "CY", "UB", "UY", "BC", "QT"),
  isUnmappedQuery = NA, yieldSize = 1e+06)
}
\arguments{
\item{reference}{Path to the reference GTF file. The TxDb object of the GTF
file will be generated and saved in current working directory with
".sqlite" suffix. Must use the same GTF file used in running the Cellranger
 pipeline.}

\item{bam}{the path to input BAM files.}

\item{cores}{Number of cores used for parallelization. Default is
\code{max(1, parallel::detectCores() - 2)}, i.e. the number of available
cores minus 2.}

\item{verbose}{Poolean indicating whether to print log messages. Useful for
debugging. Default to \strong{FALSE}.}
}
\value{
Total reads, number of aligned reads, reads aligned to an gene.
}
\description{
Read BAM file generated by Cellranger pipeline and plot QC metrics including
 total reads, number of aligned reads, reads aligned to an gene.
}
