% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot publication theme
Modify theme of ggplots. Make plots look better. Adapted from Koundinya Desiraju. https://rpubs.com/Koundy/71792}
\usage{
theme_Publication(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Default 12.}

\item{base_family}{Default \emph{sans}.}
}
\description{
ggplot publication theme
Modify theme of ggplots. Make plots look better. Adapted from Koundinya Desiraju. https://rpubs.com/Koundy/71792
}
