% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validCb}
\alias{validCb}
\title{Cell barcode whitelist (737K-august-2016.txt)}
\format{
A \code{data.table} object.
}
\usage{
validCb
}
\description{
A barcode whitelist is the list of all known barcode sequences that have
 been included in the assay kit and are available during library preparation.
 There are roughly 737,000 cell barcodes in the whitelist
 (737K-august-2016.txt) for Cell Ranger's
 Single Cell 3' and V(D)J applications.
}
\keyword{datasets}
