% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align.rsubread}
\alias{align.rsubread}
\title{A wrapper to \code{Rsubread} read alignment}
\usage{
align.rsubread(fastq.paths, index, format = "BAM", out.dir = "./Alignment",
  cores = max(1, parallel::detectCores()/2), threads = 1,
  summary.prefix = "alignment", overwrite = FALSE, verbose = FALSE,
  logfile.prefix = format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S"))
}
\arguments{
\item{fastq.paths}{A vcector containing the paths to input fastq files.}

\item{index}{Path to the \code{Rsubread} index of reference sequences. For generation of Rsubread indices, please refer to \code{buildindex} function in \code{Rsubread} package.}

\item{format}{Format of sequence alignment results. \strong{"BAM"} or \strong{"SAM"}. Default is \strong{"BAM"}.}

\item{out.dir}{Output directory for alignment results. Sequence alignment maps will be stored in folders in this directory, respectively. \strong{Make sure the folder is empty.} Default is \code{"../Alignment"}.}

\item{cores}{Number of cores used for parallelization. Default is \code{max(1, parallel::detectCores() / 2)}.}

\item{threads}{Number of threads/CPUs used for mapping for each core. Refer to \code{align} function in \code{Rsubread} for details. Default is \strong{1}.}

\item{summary.prefix}{Prefix for alignment summary file. Default is \code{"alignment"}.}

\item{overwrite}{Overwrite the output directory. Default is \strong{FALSE}.}

\item{verbose}{Print log messages. Useful for debugging. Default to \strong{FALSE}.}

\item{logfile.prefix}{Prefix for log file. Default is current date and time in the format of \code{format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S")}.}
}
\value{
A character vector of the paths to output alignment files.
}
\description{
Align cell specific reads to reference genome and write sequence alignment results to output directory. A wrapper to the \code{align} function in \code{Rsubread} package. For details please refer to \code{Rsubread} manual.
}
