% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodeExample}
\alias{barcodeExample}
\title{A vector of example cell barcodes.}
\format{A vector of cell barcode sequences. Cell barcodes for this plate are of length 8 in the example dataset (van den Brink, et al.).}
\usage{
barcodeExample
}
\description{
A vector containing 96 predefined cell barcodes which will be used for demultiplexing the example FASTQ files.
}
\keyword{datasets}
