% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{collectQC}
\alias{collectQC}
\title{Collect QC metrics}
\usage{
collectQC(de, al, co, geneAnnotation = NA)
}
\arguments{
\item{de}{Demultiplex result. Table returned from \code{demultiplex} function.}

\item{al}{Alignment result. Table returned from \code{alignRsubread} function.}

\item{co}{Count matrix. Table returned from \code{countUmi} function.}

\item{geneAnnotation}{Gene information table generated from parsing GTF file. Must contain \emph{gene_id}, \emph{gene_biotype}, and \emph{seqid} columns.}
}
\value{
QC metrics table
}
\description{
Collect QC metrics from demultiplexing, alignment, and counting results. Return a \code{data.table} object containing reads, transcripts, and genes information.
}
