% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scruff_functions.R
\name{collectqc}
\alias{collectqc}
\title{Collect QC metrics}
\usage{
collectqc(de, al, co, gene.annot = NA)
}
\arguments{
\item{de}{Demultiplex result. Table returned from \code{demultiplex} function.}

\item{al}{Alignment result. Table returned from \code{align.rsubread} function.}

\item{co}{Count matrix. Table returned from \code{count.umi} function.}

\item{gene.annot}{Gene information table generated from parsing GTF file. Must contain \emph{gene_id}, \emph{gene_biotype}, and \emph{seqid} columns.}
}
\value{
QC metrics table
}
\description{
Collect QC metrics from demultiplexing, alignment, and counting results. Return a \code{data.table} object containing reads, transcripts, and genes information.
}
