% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Demultiplex cell barcodes and assign cell specific reads}
\usage{
demultiplex(fastq, bc, bc.pos = c(6, 11), umi.pos = c(1, 5), keep = 50,
  bc.qual = 10, out.dir = "../Demultiplex",
  summary.prefix = "demultiplex", overwrite = FALSE, cores = max(1,
  parallel::detectCores() - 1), verbose = FALSE,
  logfile.prefix = format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S"))
}
\arguments{
\item{fastq}{Can be in one of the following formats: \enumerate{
\item An annotation data table or data frame that contains information about input fastq files. For example, please see \code{?exampleannot}.
\item The directory to fastq files. }}

\item{bc}{A vector of cell barcodes determined from experimental design. For example, please see \code{?examplebc}.}

\item{bc.pos}{An integer vector of length 2 consisting of the start and end index of barcodes (one-based numbering). Default is \code{c(6, 11)}.}

\item{umi.pos}{An integer vector of length 2 consisting of the start and end index of umi sequences (one-based numbering). Default is \code{c(1, 5)}.}

\item{keep}{Read length or number of nucleotides to keep for read that contains transcript sequence information. Longer reads will be clipped at 3' end. Default is \strong{50}.}

\item{bc.qual}{Minimal Phred quality score acceptable for barcode and umi sequences. Phread quality scores are calculated for each nucleotide in the sequences. Sequences with at least one score lower than this will be filtered out. Default is \strong{10}.}

\item{out.dir}{Output directory for demultiplexing results. Demultiplexed fastq files will be stored in folders in this directory, respectively. \strong{Make sure the folder is empty.} Default is \code{"../Demultiplex"}.}

\item{summary.prefix}{Prefix for summary files. Default is \code{"demultiplex"}.}

\item{overwrite}{Whether to overwrite the output directory or not. Default is \strong{FALSE}.}

\item{cores}{Number of cores used for parallelization. Default is \code{max(1, parallel::detectCores() - 1)}.}

\item{verbose}{Print log messages. Useful for debugging. Default to \strong{FALSE}.}

\item{logfile.prefix}{Prefix for log file. Default is current date and time in the format of \code{format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S")}.}
}
\value{
Demultiplexed annotation \code{data.table}.
}
\description{
Demultiplex fastq files and write cell specific reads in compressed fastq format to output directory
}
