% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleannot}
\alias{exampleannot}
\title{Example annotation table}
\format{A \code{data.table} with 5 columns:
\describe{
\item{\code{project}}{Project name. It is advised to keep project names identical across samples.}
\item{\code{sample}}{Sample name}
\item{\code{lane}}{Flow cell lane number. If fastq files from multiple lanes are concatenated after sequencing, any placeholder would be sufficient, e.g. "L001".}
\item{\code{read1_path}}{Path to read1 fastq file}
\item{\code{read2_path}}{Path to read2 fastq file}
}}
\usage{
exampleannot
}
\description{
A \code{data.table} object of the annotation table for example data.
}
\keyword{datasets}
