% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplebc}
\alias{examplebc}
\title{Example barcodes.}
\format{A vector of barcode sequences.}
\usage{
examplebc
}
\description{
96 barcodes for demultiplexing example fastq files.
}
\keyword{datasets}
