% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplefastq}
\alias{examplefastq}
\title{Example fastq \code{ShortReadQ} objects}
\format{A list of 4 \code{ShortReadQ} \code{S4} objects each containing 20,000 reads.}
\usage{
examplefastq
}
\description{
A list of \code{ShortReadQ} objects of example fastq reads (van den Brink, et al.). Contains read 1 (R1, UMI + cell barcodes) and read 2 (R2, genomic sequence) reads. Read library was generated using CEL-SEQ protocol.
}
\keyword{datasets}
