% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcplot}
\alias{qcplot}
\title{Visualize data quality}
\usage{
qcplot(qc.dt)
}
\arguments{
\item{qc.dt}{An QC metrics table returnd from \code{collect.qc} function.}
}
\value{
A list of \code{grobs} objects
}
\description{
Visualize data quality from QC metrics table and return a list of \code{grobs} objects.
}
