% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuff.R
\name{scuff}
\alias{scuff}
\title{Run scuff pipeline}
\usage{
scuff(fastq, barcodes, index, annot, length, umi.pos, bc.pos, bc.qual = 10,
  out = "./", overwrite = TRUE, ncore = 1, nthreads = 1)
}
\arguments{
\item{fastq}{An annotation data table or data frame of input fastq files.}

\item{index}{Rsubread index for reference sequences.}

\item{annot}{Directory of the gene annotation file.}

\item{umi.pos}{Start and end index number of umi sequences.}

\item{bc.pos}{Start and end index number of barcodes.}

\item{bc.qual}{Minimal acceptable quality score for barcode sequence.}

\item{out}{Output directory.}

\item{overwrite}{Overwrite the output directory. Default is TRUE.}

\item{ncore}{Number of cores to use.}

\item{bc}{A vector of cell barcodes known from experimental design.}

\item{clip.at}{Read length or number of nucleotides to keep. Longer reads will be clipped at 3' end.}

\item{threads}{Number of threads to run for each core. Default is \strong{16}.}
}
\description{
scuff runs the whole pipeline including demultiplex, align, and count_umi. Write tables of count matrices in output directory.
}
