% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuff.R
\name{scuff}
\alias{scuff}
\title{Run scuff pipeline}
\usage{
scuff(fastq, bc, index, features, bc.pos = c(6, 11), umi.pos = c(1, 5),
  keep = 50, min.qual = 10, alignment.file.format = "BAM",
  demultiplex.out.dir = "../Demultiplex",
  alignment.out.dir = "../Alignment", umi.out.dir = "../Count",
  demultiplex.summary.prefix = "demultiplex",
  alignment.summary.prefix = "alignment", expr.prefix = "count",
  logfile.prefix = format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S"),
  overwrite = FALSE, verbose = FALSE, cores = max(1,
  parallel::detectCores() - 1), threads = 1)
}
\arguments{
\item{fastq}{Can be in one of the following formats: \enumerate{
\item An annotation data table or data frame that contains information about input fastq files. For example, see \code{?exampleannot}.
\item The directory to fastq files. }}

\item{bc}{A vector of cell barcodes determined from experimental design. For example, see \code{?examplebc}.}

\item{index}{Directory to the \code{Rsubread} index of reference sequences. For generation of Rsubread indices, please refer to \code{buildindex} function in \code{Rsubread} package.}

\item{features}{Directory to the gtf reference file. For generation of TxDb objects from gtf files, please refer to \code{makeTxDbFromGFF} function in \code{GenomicFeatures} package.}

\item{bc.pos}{An integer vector of length 2 consisting of the start and end index of barcodes (one-based numbering). Default is \code{c(6, 11)}.}

\item{umi.pos}{An integer vector of length 2 consisting of the start and end index of umi sequences (one-based numbering). Default is \code{c(1, 5)}.}

\item{keep}{Read length or number of nucleotides to keep for read that contains transcript sequence information. Longer reads will be clipped at 3' end. Default is \strong{50}.}

\item{min.qual}{Minimal acceptable Phred quality score for barcode and umi sequences. Phread quality scores are calculated for each nucleotide in the sequence. Sequences with at least one nucleotide with score lower than this will be filtered out. Default is \strong{10}.}

\item{alignment.file.format}{Format of sequence alignment results. \strong{"BAM"} or \strong{"SAM"}. Default is \strong{"BAM"}.}

\item{demultiplex.out.dir}{Output directory for demultiplexing results. Demultiplexed fastq files will be stored in folders in this directory, respectively. \strong{Make sure the folder is empty.} Default is \code{"../Demultiplex"}.}

\item{alignment.out.dir}{Output directory for alignment results. Sequence alignment maps will be stored in folders in this directory, respectively. \strong{Make sure the folder is empty.} Default is \code{"../Alignment"}.}

\item{umi.out.dir}{Output directory for UMI counting results. Expression table will be stored in this directory. Default is \code{"../Count"}.}

\item{demultiplex.summary.prefix}{Prefix for demultiplex summary file. Default is \code{"demultiplex"}.}

\item{alignment.summary.prefix}{Prefix for alignment summary file. Default is \code{"alignment"}.}

\item{expr.prefix}{Prefix for expression table filename. Default is \code{"countUMI"}.}

\item{logfile.prefix}{Prefix for log file. Default is current date and time in the format of \code{format(Sys.time(), "\%Y\%m\%d_\%H\%M\%S")}.}

\item{overwrite}{Overwrite the output directory. Default is \strong{FALSE}.}

\item{verbose}{Print log messages. Useful for debugging. Default to \strong{FALSE}.}

\item{cores}{Number of cores used for parallelization. Default is \code{max(1, parallel::detectCores() - 1)}.}

\item{threads}{Number of threads/CPUs used for mapping for each core. Refer to \code{align} function in \code{Rsubread} for details. Default is \strong{1}.}
}
\description{
Run the \code{scuff} pipeline including \code{demultiplex}, \code{align.rsubread}, and \code{count.umi}. Write expression table in output directory.
}
