% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scruffFuncitons.R
\name{stepping}
\alias{stepping}
\title{Visualize aligned reads}
\usage{
stepping(bamGA, chr = "1", start = 1, end = max(BiocGenerics::end(bamGA)),
  legend = FALSE)
}
\arguments{
\item{bamGA}{A GenomicAlignment object}

\item{chr}{Chromosome. Integer or "X", "Y", "MT".}

\item{start}{Genomic coordinate of the start position.}

\item{end}{Genomic coordinate of the end position.}

\item{legend}{Show legend. Default is FALSE.}
}
\value{
A ggplot object of aligned reads
}
\description{
Visualize read alignments for UMI tagged single cell RNA-sequencing data. Arrow represents orientation of alignment. Reads are colored by their UMI and sorted by their start positions and UMI.
}
