% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scruff_functions.R
\name{stepping}
\alias{stepping}
\title{Visualize stepping levels}
\usage{
stepping(bamGA, chr = "1", start = 1, end = max(BiocGenerics::end(bamGA)))
}
\arguments{
\item{bamGA}{A GenomicAlignment object load a toy "BAM" file}

\item{chr}{Chromosome name or vector of the chromosome names need to check}

\item{start}{A integer of start position which need to check}

\item{end}{A integer of end position which need to check}
}
\value{
A plot of stepping levels
}
\description{
Calculate the stepping levels for single cell RNA-sequencing data.Visualize the data using ggplot2 and ggbio
}
